+++ src/main/java/refinedstorage/inventory/ItemHandlerGridFilterInGrid.java
@@ -2,13 +2,15 @@ package refinedstorage.inventory;
 
 import net.minecraft.item.ItemStack;
 import refinedstorage.RefinedStorageItems;
+import refinedstorage.gui.grid.GridFilteredItem;
+import refinedstorage.item.ItemGridFilter;
 
 import java.util.List;
 
 public class ItemHandlerGridFilterInGrid extends ItemHandlerBasic {
-    private List<ItemStack> filteredItems;
+    private List<GridFilteredItem> filteredItems;
 
-    public ItemHandlerGridFilterInGrid(List<ItemStack> filteredItems) {
+    public ItemHandlerGridFilterInGrid(List<GridFilteredItem> filteredItems) {
         super(4, new ItemValidatorBasic(RefinedStorageItems.GRID_FILTER));
 
         this.filteredItems = filteredItems;
@@ -21,14 +23,16 @@ public class ItemHandlerGridFilterInGrid extends ItemHandlerBasic {
         filteredItems.clear();
 
         for (int i = 0; i < getSlots(); ++i) {
-            ItemStack stack = getStackInSlot(i);
+            ItemStack filter = getStackInSlot(i);
 
-            if (stack != null) {
-                ItemHandlerGridFilter items = new ItemHandlerGridFilter(stack);
+            if (filter != null) {
+                int compare = ItemGridFilter.getCompare(filter);
+
+                ItemHandlerGridFilter items = new ItemHandlerGridFilter(filter);
 
                 for (ItemStack item : items.getFilteredItems()) {
                     if (item != null) {
-                        filteredItems.add(item);
+                        filteredItems.add(new GridFilteredItem(item, compare));
                     }
                 }
             }