+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 **Features**
 - Added German translation by ChillUpX
+- Each Grid Filter can now only filter 9 items
+- Each Grid allows 4 Grid Filters instead of 1
+- Grid Filters can now be configured to compare on NBT and/ or damage
 
 ### 0.8.15
 **Bugfixes**